refactor: 매직 넘버 삭제
