+++ src/main/java/constants/ErrorMessage.java
@@ -1,4 +1,4 @@
-package message;
+package constants;
 
 public enum ErrorMessage {
     NUMBER_FORMAT("양의 정수만 입력할 수 있습니다."),
