+++ src/main/java/constants/LottoConstants.java
@@ -1,0 +1,7 @@
+package constants;
+
+public class LottoConstants {
+    public static final int LOTTO_PRICE = 1000;
+    public static final int LOTTO_NUMBERS_SIZE_THRESHOLD = 6;
+    public static final int LOTTO_NUMBER_MAX_THRESHOLD = 45;
+}
