+++ src/main/java/domain/numbergenerator/RandomNumberGenerator.java
@@ -1,5 +1,6 @@
 package domain.numbergenerator;
 
+import constants.LottoConstants;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -8,8 +9,8 @@ public class RandomNumberGenerator implements NumberGenerator {
     @Override
     public List<Integer> generateNumber() {
         Set<Integer> selectedNumber = new HashSet<>();
-        while (selectedNumber.size() != 6) {
-            int pickedNumber = (int) (Math.random() * 45) + 1;
+        while (selectedNumber.size() != LottoConstants.LOTTO_NUMBERS_SIZE_THRESHOLD) {
+            int pickedNumber = (int) (Math.random() * LottoConstants.LOTTO_NUMBER_MAX_THRESHOLD) + 1;
             selectedNumber.add(pickedNumber);
         }
