+++ src/test/java/domain/numbergenerator/RandomNumberGeneratorTest.java
@@ -1,6 +1,7 @@
 package domain.numbergenerator;
 
 
+import constants.LottoConstants;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -17,6 +18,6 @@ class RandomNumberGeneratorTest {
         List<Integer> selectedNumber = randomNumberGenerator.generateNumber();
 
         // then
-        Assertions.assertThat(selectedNumber.stream().distinct()).hasSize(6);
+        Assertions.assertThat(selectedNumber.stream().distinct()).hasSize(LottoConstants.LOTTO_NUMBERS_SIZE_THRESHOLD);
     }
 }
